clean up handle
